requirements.txt
# tensorboard very slow to install on MacOS, maybe delete
tensorboard
# wandb
pytest
numpy

# for loading dataset examples
pandas
ucimlrepo

# for Mathematica comms
pyzmq

# for line profiling
line_profiler

pytorch
torchvision
pillow
six
wandb
scipy
setuptools